get rid of pkg/hyperkube
