+++ cmd/hyperkube/hyperkube_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package hyperkube
+package main
 
 import (
 	"bytes"
