+++ cmd/hyperkube/kube-scheduler.go
@@ -17,19 +17,18 @@ limitations under the License.
 package main
 
 import (
-	"github.com/GoogleCloudPlatform/kubernetes/pkg/hyperkube"
 	scheduler "github.com/GoogleCloudPlatform/kubernetes/plugin/cmd/kube-scheduler/app"
 )
 
 // NewScheduler creates a new hyperkube Server object that includes the
 // description and flags.
-func NewScheduler() *hyperkube.Server {
+func NewScheduler() *Server {
 	s := scheduler.NewSchedulerServer()
 
-	hks := hyperkube.Server{
+	hks := Server{
 		SimpleUsage: "scheduler",
 		Long:        "Implements a Kubernetes scheduler.  This will assign pods to kubelets based on capacity and constraints.",
-		Run: func(_ *hyperkube.Server, args []string) error {
+		Run: func(_ *Server, args []string) error {
 			return s.Run(args)
 		},
 	}
